cryptohome: Make SanitizeUserName() compatible with cryptohome
SanitizeUserName() does not produce strings which are compatible
with cryptohome.  This is a known issue that was supposed
to be resolved before this code went into common use.

This change just downcases the sanitized username to match
cryptohome. This _should_ only impact ephemeral mountpoints in
/home/user and /home/root which should be empty and cleaned up.
(They should be changed to tmpfs mountpoints.)

TEST=emerged, output matches cryptohome (lowercase), TODO still - boot test, BVT run
     Tested that IsSanitizedUserName() returns true for the new lowercase output.
BUG=chromium-os:38444

Change-Id: I8be838106fa9df9e51ac31d9d6fa5887f7203a45
Reviewed-on: https://gerrit.chromium.org/gerrit/42520
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>