simplify usage
There is no reason to strdup the string (even with the POSIX basename),
so drop that.

Further, we want to use the GNU variant which doesn't modify its argument,
so drop the libgen.h include.  The string.h header will provide the sane
variant.

BUG=None
TEST=built & ran bootstat

Change-Id: I9cd55e0ce064c0a04e9f13c503b334cb98b065c1
Reviewed-on: https://chromium-review.googlesource.com/184626
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>