libchromeos: Add CallMethod() to help with unit testing.
Added CallMethod() helper method and swept current unit test
to use it instead of rolling out their own. This method will also
be used in more unit tests in the future.

To ensure it is not included in the shipping bits, it was made
inline, so unless someone calls it, it will not produce any
executable bits.

BUG=chromium:374864
TEST=FEATURES=test emerge-link libchromeos

Change-Id: I39097d8c14f8e653ee656df25159ca49facd483b
Reviewed-on: https://chromium-review.googlesource.com/213265
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>