libchromeos: Implement D-Bus signal registration
Added a way to register a signal method on a DBusInterface.
When a signal is registered, a callback proxy object is returned
that can be able to dispatch a D-Bus signal as if calling a native
C++ method.

Switched current signal dispatching in DBusObject-based objects
to use the new signal mechanism.

BUG=None
TEST=FEATURES=test emerge-link libchromeos peerd buffet

Change-Id: Icbf9da90fa0a6952a903cf6f37add19f84b88fbf
Reviewed-on: https://chromium-review.googlesource.com/214553
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>