libchromeos: policy: Fail GetHttpDownloadsEnabled() when field is not set.
Update the GetHttpDownloadsEnabled method to fail (return false) when
the field is not set. Up until now, it will fail if the
auto_update_settings proto wasn't set at all, but it will return the
default value specified in the .proto defining it is the proto is set
but the field is not.

This patch makes this function fail if the proto is not set or the field
is not present. This makes update_engine default to 'true' when the field
is not present instead of the default value present in the .proto, which
at the moment is false.

The policy_all file was updated to include the http_downloads_enabled
field with the "false" value..

BUG=chromium:407917
TEST=`FEATURES=test emerge-link libchromeos`

Change-Id: I7ae02bbbb00a9377483068ed23eaff1be1be4df4
Reviewed-on: https://chromium-review.googlesource.com/214657
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>