libchromeos: Fix potential reference to temporary in Any class
On some compilers/configurations, Any::TryGet() is not inlined and
the default value parameter is not rolled into the call site, causing
returning a reference to a temporary variable.

We will pay a small performance penalty for this safe behavior,
since we are now returning the copy of the contained data by value
now rather than reference.

BUG=None
TEST=FEATURES=test emerge-link libchromeos

Change-Id: I4cba31ba5a46341f10ab055fdb9cd88ef4596615
Reviewed-on: https://chromium-review.googlesource.com/219133
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>