libchromeos: Validate file descriptors when obtaining over D-Bus
With automatic data deserialization mechanisms in libchromeos,
the data received from D-Bus message buffer is passed to a callback
as const reference. This means that if FileDescriptor is one of
the parameters, it will be const by the time it arrives to the
handler and there is no way of validating the file descriptor value.
Added a call to FileDescriptor::CheckValidity() in PopValueFromReader
to make sure we do it before the object turns constant.

BUG=None
TEST=FEATURE=test emerge-link libchromeos

Change-Id: I988f3729767b2dd8f094ce332fb4fef4eff36653
Reviewed-on: https://chromium-review.googlesource.com/219424
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>