libchromeos: ExtractMethodCallResults returns error when D-Bus method fails
ExtractMethodCallResults() would not detect an error response from
a D-Bus method call and if the caller expects the method to return
a string, the D-Bus error message would be sent to the caller as
if it were the method's return value.

Fixed ExtractMethodCallResults() to detect ErrorResponse message
from D-Bus and use the error details to return through the |error|
object.

Also added AddDBusError method to complement GetDBusError which
makes it possible to serialize a chain of errors on the D-Bus
method handler side and then reconstruct it on the caller side.

BUG=None
TEST=FEATURES=test emerge-link libchromeos

Change-Id: I1c302a3e5616d3d3e3b0225fd4ca7ef09fd4faab
Reviewed-on: https://chromium-review.googlesource.com/227510
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>