libchromeos: Allow client to set response stream
Previously the curl connection always used a MemoryStream to write
the http response to. This change gives the user more flexibility to
specify other types of streams to send the response to, for example
maybe you would like to be able to stream it to a file, or /dev/null, or
some other type of stream.

BUG=brillo:1174
TEST=FEATURES=`test emerge-amd64-generic libchromeos` + manually tested

Change-Id: Iaa9edbffe5d76f71f56a7e4766c17bd641afb653
Reviewed-on: https://chromium-review.googlesource.com/275640
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Nathan Bullock <nathanbullock@google.com>
Tested-by: Nathan Bullock <nathanbullock@google.com>