Fix Any::Buffer.Store_Objects unit test
Any normally embeds small data objects (with the size of up to
double). This test verifies that larger objects are stored externally.

However 10 bytes is not large enough and on some platform due to data
alignment/padding a bit larger data than that of double can still be
stored inside Any itself. Use 20 bytes to guarantee that Any cannot
contain the buffer of that size inside.

Change-Id: Iceaa0efbc74570f17d2cc70aa8d6537cc8bd2be5