libbrillo: Don't delete and add the binder fd in BaseMessageLoop.
The binder driver polling mechanism doesn't work well with epoll. In
particular, you can't add the binder file descriptor to epoll (with
EPOLL_CTL_ADD) when there's work to do in the binder driver, then
wait for it when there isn't any more work to do.

To avoid this situation, we only add the binder file descriptor at the
beginning and never try to remove and add it back when running
callbacks for it.

Bug: 26524111
TEST=unittests for parsing.
TEST=Tested with an example service, strace shows the binder fd is not added/removed each time.
Change-Id: I11d6af5593fa78631f80e9ba09d933f4b8ef70df

Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: cf6e947ef9c02c64614af966c1c5c5ca38c706e6