Add DBusConnection helper class to use instead of DBusDaemon.
This CL moves the logic in DBusDaemon to a helper class that can be
included in another class inheriting from Daemon without needed to use
multiple inheritance nor forcing a DBus connection to succeed.
DBusDaemon was modified to re-use this class without changing its
behaviour to be compatible with existing code.

Bug: 26690086
TEST=mm; Update update_engine to use this class and tested it on edison.

Change-Id: If05a5582158552b6aeebf86dbde2aaa161f5a29f