libbrillo: policy: Expose allow_kiosk_app_control_chrome_version policy
Expose allow_kiosk_app_control_chrome_version policy. Kiosk apps
can declare a "required_platform_version" manifest key. The policy
controls whether the value of the key should be used.  When the
policy is set to true, the value is used as target version prefix
in the AU ping. When the policy is not configured or set to false,
the manifest key is ignored and AU works as normal. The policy
defaults to false.

BUG=chromium:577765
TEST=libpolicy_unittest
Change-Id: I8889f27497e806cdc71e7da743474d402a89efd5

Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 5419bb0b07df06e72074d93e518dc541cdd41be3