Call MessageLoop::current() to detect if a message loop was created.
MessageLoopForIO::current() and MessageLoop::current() are very similar
functions, except that the former will crash when no message loop was
created *and* we are compiling with debug check due to a pointer
dereference in a DCHECK(). While we can make this failure explicit in
libchrome, this patch avoids the issue by calling
MessageLoop::current() instead.

Bug: None
TEST=FEATURES=test emerge-link libbrillo metrics

Change-Id: I5f446ac55b6459d460fffdc47316cdf2c588c1d8