libbrillo: Make dbus code optional
The dbus code is pretty self-contained and the Android.mk already allows
you to skip the dbus library if you don't want it.  Also make it
optional in the gyp file based on a USE flag.  The USE flag is set / not
set with a corresponding change in the ebuild.

BUG=none
TEST=emerge-$BOARD libbrillo with and without the USE flag set and see
that the dbus library is present / not present.
CQ-DEPEND=CL:455269

Change-Id: Ibfb46ebe400973dad9b3f2e3ac072b27cc86edcb
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/455548
Commit-Ready: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 989065b1cc9f2bbb2d41e749fac1a7b720d0391e