Convert most libbrillo targets to soong.
The "libbrillo-minijail" target is left outside of the conversion since
it depends on the minijail C project.

Converted using the androidmk.go tool and manually merging common
host/target sections.

As a note, the inclusion of the local-generated-sources-dir in the .mk
was not needed anymore after the removal of the DBus helpers.

Bug: None
Test: `make checkbuild`
Change-Id: Ie3c7aa130f16e8d71ae72937e47efdafc7cde1a6