libbrillo: use ScopedTempDir::GetPath
Upstream versions of libchrome have gotten rid of ScopedTempDir's
inline path method because they wanted to introduce a DCHECK that
helps make sure people create the temp directory before using it.
To avoid introducing a header dependency on the logging header
they took this method out of the header so it wouldn't be inlined.

BUG=b:37434548
TEST=unit tests

Change-Id: Ic83227a7ddbbad32a1df1f9554d29470c1a2dd42
Reviewed-on: https://chromium-review.googlesource.com/905784
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>