libbrillo: use bound arguments rather than capture-by-ref
In future libchrome we will have the ability to bind lambdas but
it's considered unacceptable to bind lambdas that carry an
environment. To prepare for this, explicitly state the environment
as bound bare pointers to local variables.

BUG=b:37434548
TEST=unit tests

Change-Id: Id6cd318077de9b5ad253731a107d2cb418323de6
Reviewed-on: https://chromium-review.googlesource.com/1003892
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 8151e43f513a6c58882f1a10c368cb413ec6a590