debugd: Add LsbRelease contents
Partners will frequently share debug logs to provide context when filing
a bug or diagnosing a problem. Feedback reports have this information,
but it's not stored in any of the files that are packaged up with debug
logs.

This commit adds version information to debugd output so that it is
present in debug logs.

A subsequent change will be made in the chrome source. That change will
stop chrome from fetching the lsb release contents on its own, which is
ok to do so because debugd will be fetching the lsb contents.
Here is the link to the subsequent chrome change:
https://chromium-review.googlesource.com/c/chromium/src/+/1217348

BUG=chromium:880851
TEST=Manually checked chrome://system

Change-Id: I37e6317754df42266d36fcbfac7a43bb68c23ac0
Reviewed-on: https://chromium-review.googlesource.com/1217083
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Kartik Hegde <khegde@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 1c4b97b3a5743582f8ee096830b7cf0ef265f2d4