libbrillo: Conditionally quote boundary field in Content-Type
While technically it is not incorrect to do this, there are
implementations of HTTP servers that don't handle this properly.
A specific example is the server that handles uploads of crash reports.
I've contacted them about fixing it on their end as well...but we might
as well change it on ours too.

BUG=None
TEST=Verified uploads to crash server work with this change

Change-Id: I143a0823e8b22e67ad5b54e8e86c0378ad227e10
Reviewed-on: https://chromium-review.googlesource.com/1584418
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Tim Zheng <timzheng@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 33b6f16e53995b912b8096c4068eafc3753de730