libbrillo: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes this code
clearer and easier to modify, since if a case is added to this `switch`
later, the user should get a -Werror'ed warning if they forget to add a
`return`.

BUG=chromium:1066998
TEST=CQ

Change-Id: If0be83daa6504854c377c825cf345c2c4d39837b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2132836
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Cr-Mirrored-From: https://chromium.googlesource.com/chromiumos/platform2
Cr-Mirrored-Commit: 6d0b86eeb049f098a1735fff8a06b62f9689a3c8